+++ oc_lettings_site/settings.py
@@ -22,6 +22,7 @@ ALLOWED_HOSTS = []
 
 INSTALLED_APPS = [
     'oc_lettings_site.apps.OCLettingsSiteConfig',
+    'lettings',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
