+++ oc_lettings_site/urls.py
@@ -1,12 +1,12 @@
 from django.contrib import admin
-from django.urls import path
+from django.urls import include, path
 
 from . import views
 
+
 urlpatterns = [
     path('', views.index, name='index'),
-    path('lettings/', views.lettings_index, name='lettings_index'),
-    path('lettings/<int:letting_id>/', views.letting, name='letting'),
+    path('lettings/', include('lettings.urls')),
     path('profiles/', views.profiles_index, name='profiles_index'),
     path('profiles/<str:username>/', views.profile, name='profile'),
     path('admin/', admin.site.urls),
